feat: aleo private sync for mobile part 1
