+++ .changeset/curly-poems-ring.md
@@ -1,0 +1,7 @@
+---
+"ledger-live-desktop": minor
+"live-mobile": minor
+"@ledgerhq/live-common": minor
+---
+
+private sync for mobile Aleo part 1
