+++ apps/ledger-live-desktop/src/renderer/families/aleo/constants.ts
@@ -1,10 +1,3 @@
-export const PRIVATE_BALANCE_PLACEHOLDER = "***";
-
 export enum AleoCustomModal {
   SELF_TRANSFER = "MODAL_ALEO_SELF_TRANSFER",
 }
-
-export const MANDATORY_SYNC_POLLING_DELAY = 3000;
-
-/** Minimum time (ms) between progress state updates in the hook to avoid flooding React renders. */
-export const PROGRESS_THROTTLE_INTERVAL_MS = 500;
