--- apps/ledger-live-mobile/src/families/aleo/constants.ts
@@ -1,1 +1,0 @@
-export const PRIVATE_BALANCE_PLACEHOLDER = "***";
