+++ libs/ledger-live-common/src/families/aleo/__mocks__/currency.mock.ts
@@ -1,0 +1,3 @@
+import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets";
+
+export const aleoCurrency = getCryptoCurrencyById("aleo");
